Add Functions For Actions That Server Will Create
Add functions to handle actions that the server will create and send
them to the dispatch to be process. Because this way, multiple functions
could use the same action for their response. In this case it is for
logging in as once you have sign up you should be login automatically
and the actually login will use the same functions.