Add Flux Store To Handle Session
Add in a Flux Store to handle the session because the session needs to
set its valuables and send back an event to the view in order to display
if the login was successful or not.

A Flux Store is a mix between a model and a controller, it handles the
data which is the main state of the application, then feeds the records
to the views, in the form of an event emitter, while retrieving that
data from a server.