Add Component SmallApp, Main Component For App
Add the component SmallApp, which is the main component for the
application, this is where the header for the application and which
other components will load because of the routes. This way there is only
one component to change if there needs to be a application wide change.

What is currently happening the user initiates a view action, the
WebAPIUtils calls the server, serve responses and rasies a new aciton,
the dispatcher forwards it to the SessionStore, which updates its status
and emits a change event then catched by the SmallApp component. The
SmallApp component forwrds its state to its child which is the Header
component.